Moved dynamic table to table file
