+++ src/http/qpack/mod.rs
@@ -3,13 +3,7 @@
 #![allow(unused_imports)]
 
 pub mod table;
-use self::table::HeaderField;
-
-pub mod dyn_table;
-use self::dyn_table::DynamicTable;
-
 pub mod static_table;
-use self::static_table::StaticTable;
 
 pub mod parser;
